requirements.txt
asgiref==3.8.1
Django==5.2.3
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
